pyproject.toml
[project]
name = "csv-analysis-agent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "langchain>=1.2.10",
    "langchain-classic>=1.0.1",
    "langchain-experimental>=0.4.1",
    "langchain-openai>=1.1.10",
    "openai>=2.21.0",
    "pandas>=3.0.1",
    "python-dotenv>=1.2.1",
]
